# docs/swagger.yaml
info:
  contact: {}
paths:
  /students:
    get:
      consumes:
      - application/json
      description: Retrieve entities details
      parameters:
      - description: Registration
        in: path
        name: register
        type: integer
      produces:
      - application/json
      responses:
        "404":
          description: Not Found
      summary: Get a list of entities
      tags:
      - students
    post:
      consumes:
      - application/json
      description: Create a new entity asking all parameters of the student struct
      parameters:
      - description: Registration
        in: path
        name: register
        type: integer
      produces:
      - application/json
      responses:
        "404":
          description: Not Found
      summary: Create entity
      tags:
      - students
  /students/:id:
    delete:
      consumes:
      - application/json
      description: Delete a specific entity
      parameters:
      - description: Registration
        in: path
        name: register
        type: integer
      produces:
      - application/json
      responses:
        "404":
          description: Not Found
      summary: Delete entity
      tags:
      - students
    get:
      consumes:
      - application/json
      description: Retrive information about a specific entity
      parameters:
      - description: Registration
        in: path
        name: register
        type: integer
      produces:
      - application/json
      responses:
        "404":
          description: Not Found
      summary: Get entity by ID
      tags:
      - students
    put:
      consumes:
      - application/json
      description: Update new information about a specific entity
      parameters:
      - description: Registration
        in: path
        name: register
        type: integer
      produces:
      - application/json
      responses:
        "404":
          description: Not Found
      summary: Update entity
      tags:
      - students
swagger: "2.0"